+++ tests/Unit/Messaging/MessageTargetTest.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kreait\Firebase\Tests\Unit\Messaging;
+
+use Kreait\Firebase\Exception\InvalidArgumentException;
+use Kreait\Firebase\Messaging\MessageTarget;
+use PHPUnit\Framework\TestCase;
+
+class MessageTargetTest extends TestCase
+{
+    public function testInvalidTargetCausesError()
+    {
+        $this->expectException(InvalidArgumentException::class);
+        MessageTarget::with('foo', 'bar');
+    }
+}